Update 35.in-situ-data.md
Changed order of sections so that paragraphs are in chronological order (i.e. accumulation, ablation then surface elevation surveys). This new order matches the order in which the surveys are listed in the introductory paragraph. This was helpful for describing how the snow pits were used to calculate meltwater equivalents from the ablation stakes.

See my edits to the second introductory paragraph - how do those edits look to you?